treewide: Rework API
The API methods now operate directly on the database. Therefore, the
classes don't have a `save()` method anymore.

Also, the distinction between tracks and their performance data has
been hidden from the API user. That is, the performance data attributes
are now accessible directly through the `track` API.

This is WORK IN PROGRESS. See the TODOs.