Use std::optional where available
Either <optional> or <experimental/optional> is now required for
compilation and in the public API.  In order to seamlessly use either
one, the relevant symbols are aliased within the namespace `djinterop::stdx`.

It is expected that this is removed in favour of just using
std::optional in future once older compiler support can be removed.