API has w/form samples-per-entry rather than size
Informing the user of a desired waveform size is not the right approach,
as the waveform is still essentially meaningless without knowledge of
the scale it uses.  As such, the public API now exposes the required
number of _samples per waveform entry_, thereby creating a clear-cut
scaling factor for waveform interpretation in the context of samples.